% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{multiplesheets}
\alias{multiplesheets}
\title{readin excel and sum all sheets}
\usage{
multiplesheets(fname)
}
\arguments{
\item{fname}{}
}
\value{
sheets name list
}
\description{
readin excel and sum all sheets
}
\examples{
library(rSplitExcel)
multiplesheets(fname)
}
