% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{multiplesheets_write}
\alias{multiplesheets_write}
\title{write sheets with same title to one file}
\usage{
multiplesheets_write(fname)
}
\arguments{
\item{fname}{}
}
\value{
output excels
}
\description{
write sheets with same title to one file
}
\examples{
library(rSplitExcel)
multiplesheets_write(fname)
}
